Added metadata to Diagnostic Context and replacing Device.java call locations
